[FE] refactor : 유저 프로필 요청 API 연결
